Implement the Bayesian hierarchical PopReconstruct model

The popReconstruct model embeds the cohort component method of population projection (ccmpp) in a Bayesian hierarchical model. It reconciles ccmpp input preliminary estimates of sex-ratio at birth, fertility, mortality, migration, and a baseline population with population data while accounting for measurement error in these initial estimates.

popReconstruct_fit() fits the model using Stan or TMB.

popReconstruct_posterior_draws() produces draws from the posterior distribution.

popReconstruct_prior_draws() produces draws from the prior distribution.

popReconstruct_summarize_draws() produces summary statistics of draws from the popReconstruct model using demUtils::summarize_dt(). The summarize_cols parameter should include 'chain', 'chain_draw' and 'draw' when the model was fit with stan and just 'draw' when the model was fit with tmb.

popReconstruct_fit(
  inputs,
  data,
  hyperparameters,
  settings,
  value_col,
  software,
  ...
)

popReconstruct_posterior_draws(
  fit,
  inputs,
  settings,
  value_col,
  software,
  method_name
)

popReconstruct_prior_draws(
  inputs,
  hyperparameters,
  settings,
  value_col,
  method_name,
  chunk_size = 100
)

popReconstruct_summarize_draws(
  draws,
  summarize_cols = c("chain", "chain_draw", "draw"),
  value_col = "value",
  ...
)

Arguments

inputs
[list()]
[data.table()] for each ccmpp input. Requires 'srb', 'asfr', 'baseline', and 'survival'; and migration estimates provided as just 'net_migration' or both 'immigration' and 'emigration'. See Section: Inputs for more information on each of the required inputs.

data
[list()]
[data.table()] for each type of data input for the model (currently only 'population'). See Section: Data for more information.

hyperparameters
[list()]
List for each model component ('srb', 'asfr', ...) that itself contains a list specifying the alpha (shape) and beta (scale) hyperparameters for the variance of each model component.

settings
[list()]
Named list of settings for running the popReconstruct model with. The required settings are the same as those required for demCore::ccmpp(), see Section: Settings for these. The optional settings specific to popReconstruct are described in Section: Optional popReconstruct Settings

value_col
[character(1)]
Name of the column containing the value of interest in each of the inputs. Default is 'value'.

software
[character()]
Statistical modeling software to fit popReconstruct with. Either 'stan' or 'tmb' are available.

...
For popReconstruct_fit() additional arguments to pass to rstan::sampling() if fitting the model with 'stan' or optimx::optimx() if fitting the model with 'tmb'. For popReconstruct_summarize_draws() additional arguments to pass to demUtils::summarize_dt().

fit
[stanfit(1)] or [sdreport(1)]
Model fit object returned by popReconstruct_fit().

method_name
[character(1)]
Description to assign to column 'method' in each returned data.table.

chunk_size
[integer(1)]
number of draws to sample from the prior at once.

draws
[list()] of [data.table()]
Draws from the posterior distribution of the popReconstruct model as returned by popReconstruct_posterior_draws().

summarize_cols
[character()]
The id_cols that should be collapsed and to calculate summary statistics over.

Optional popReconstruct Settings

fixed_parameters: [character()]
Names of demCore::ccmpp() input components that should be fixed in the model, corresponds to the names of the input list. Defaults to empty if fitting the model for both sexes, if only 'female' then defaults to 'srb'.

n_draws: [character()]
The number of draws to take from the output posterior distribution. If fitting the model with stan then the number of draws is dictated by chains, iterations, burn-in, thin arguments to rstan::sampling; n_draws can then be used to subsample down.

k_years_input_name: [integer()]
The start of each calendar year interval that will be knots in the B-spline linear basis functions for the year-specific input estimates, so not relevant for 'baseline'. Defaults to the setting for 'years' which is equal to the original popReconstruct model without linear B-splines to reduce year-specific parameters.

k_ages_input_name: [integer()]
The start of each age-group interval that will be knots in the B-spline linear basis functions for the age-specific input estimates, so not relevant for 'srb'. Defaults to the setting for 'ages' (or 'ages_input_name') which is equal to the original popReconstruct model without linear B-splines to reduce age-specific parameters.

Data

population: [data.table()]

year: [integer()] year of population data.

sex: [character()] either 'female', 'male', or 'both'. If 'sexes' setting is 'female' can only have 'female' input data.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting. Age groups included must not overlap.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] population count, must be greater than zero.

popReconstruct_fit Value

If fit using 'stan' an object of class stanfit and if fit using 'tmb' an object of class sdreport. Either represents the fitted results that can be extracted using popReconstruct_posterior_draws(). Stan has helper packages that can be used to explore the model fit through shinystan, bayesplot, etc.

popReconstruct_posterior_draws and popReconstruct_prior_draws Value

popReconstruct_posterior_draws() and popReconstruct_prior_draws() return a named [list()] of [data.table()] for draws from the posterior and prior distribution respectively for each ccmpp input component along with the associated offset and spline offset parameters. Draws for the 'variance' and projected 'population' are also included.

The returned [data.table()]s will include columns related to the draw number:

draw: the draw index.

chain: if using stan to fit the model, the chain the draw came from.

chain_draw: if using stan to fit the model, the draw index for that chain.

population: [data.table()]

year: [integer()] mid-year for population estimate. Corresponds to 'years' setting plus an extra interval.

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

draw: [integer()] draw index.

value_col: [numeric()] projected population count estimates.

variance: [data.table()]

parameter: [character()] the ccmpp component the variance term corresponds to.

draw: [integer()] draw index.

value_col: [numeric()] estimated variance value..

For each ccmpp component ('srb', 'asfr', etc.) there will be three [data.table()] outputs included in the returned [list()]. These will each have the same columns as described in the Section: Inputs but with additional draw related columns as described above.

The offset parameters representing the values of the piecewise linear function at the year- and age-specific knots. These are the estimated deviations from the initial ccmpp input estimates at the exact knots. The estimated values are in the transformed modeling space (log transformed for 'srb', logit transformed for 'survival', etc.).

The spline offset values calculated by multiplying the fixed B-spline linear basis functions by the offset parameters. These are the estimated deviations from the initial ccmpp input estimates at all years and ages. The estimated values are in the transformed modeling space.

The actual ccmpp input posterior draws after combining the spline offset values with the initial ccmpp input estimates and after applying the inverse of the transformation used to model each component.

popReconstruct_summarize_draws Value

Returns a named [list()] of [data.table()] for summary statistics of the input draws. Each data.table() will have the id_cols for each component (minus the summarize_cols) plus summary statistic columns. The summary statistic columns have the same name as each function specified in summary_fun and the quantiles are named like 'q_(probs * 100)'. See [demUtils::summarize_dt()] for more information.

years: [numeric()]
The start of each calendar year interval to project in demCore::ccmpp(). Corresponds to the 'year_start' column in each of the year-specific inputs.

sexes: [character()]
The sexes being projected in demCore::ccmpp(), must be either just 'female', or 'female' and 'male'. Corresponds to the 'sex' column in each of the sex-specific inputs.

ages: [numeric()]
The ages being projected in demCore::ccmpp(). Corresponds to the 'age_start' column in each of the standard age-specific inputs.

ages_survival: [numeric()]
The ages for which survivorship ratio estimates are available, includes one extra age group compared to the 'ages' setting. Corresponds to the 'age_start' column in the 'survival' [data.table()] input.

ages_asfr: [numeric()]
The assumed female reproductive ages, subset of 'ages'. Corresponds to the 'age_start' column in the age-specific fertility rate (asfr) [data.table()] input.

Inputs

srb: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

value_col: [numeric()] sex-ratio at birth estimates, must be greater than zero.

asfr: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages_asfr' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual age-specific fertility rate estimates, must be greater than zero.

baseline: [data.table()]

year: [integer()] mid-year for population estimate. Corresponds to 'years' setting.

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] baseline year population count estimates, must be greater than zero.

survival: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] survivorship ratio estimates, must be greater than zero and less than one.

net_migration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual net-migration proportion estimates.

immigration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual immigration proportion estimates, must be greater than zero.

emigration: [data.table()]

year_start: [integer()] start of the calendar year interval (inclusive). Corresponds to 'years' setting.

year_end: [integer()] end of the calendar year interval (exclusive).

sex: [character()] either 'female' or 'male'. Corresponds to 'sexes' setting.

age_start: [integer()] start of the age group (inclusive). Corresponds to 'ages' setting.

age_end: [integer()] end of the age group (exclusive).

value_col: [numeric()] annual emigration proportion estimates, must be greater than zero.

References

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2013. “Reconstructing Past Populations With Uncertainty From Fragmentary Data.” Journal of the American Statistical Association 108 (501): 96–110. https://doi.org/10.1080/01621459.2012.737729.

popReconstruct R Package

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal Statistical Society. Series A: Statistics in Society 178 (4): 977–1007. https://doi.org/10.1111/rssa.12104.

markalava/Bayesian-Reconstruction github repo

See also

Examples

# specify settings for this example
settings = list(
  years = seq(1960, 2000, 5),
  sexes = c("female"),
  ages = seq(0, 80, 5),
  ages_survival = seq(0, 85, 5),
  ages_asfr = seq(15, 45, 5),

  n_draws = 10
)

# hyperparameters for the variance prior distribution, represents measurement error
hyperparameters <- list(asfr = list(alpha = 1, beta = 0.0109),
                        population = list(alpha = 1, beta = 0.0109),
                        survival = list(alpha = 1, beta = 0.0109),
                        net_migration = list(alpha = 1, beta = 0.0436))

fit_stan <- popMethods::popReconstruct_fit(
  inputs = demCore::burkina_faso_initial_estimates,
  data = demCore::burkina_faso_data,
  hyperparameters = hyperparameters,
  settings = settings,
  value_col = "value",
  software = "stan",
  chains = 1,
  iter = 200,
  warmup = 100,
  thin = 2
)

#> 
#> SAMPLING FOR MODEL 'popReconstruct' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.001343 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 13.43 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 15
#> Chain 1:            adapt_window = 75
#> Chain 1:            term_buffer = 10
#> Chain 1: 
#> Chain 1: Iteration:   1 / 200 [  0%]  (Warmup)
#> Chain 1: Iteration:  20 / 200 [ 10%]  (Warmup)
#> Chain 1: Iteration:  40 / 200 [ 20%]  (Warmup)
#> Chain 1: Iteration:  60 / 200 [ 30%]  (Warmup)
#> Chain 1: Iteration:  80 / 200 [ 40%]  (Warmup)
#> Chain 1: Iteration: 100 / 200 [ 50%]  (Warmup)
#> Chain 1: Iteration: 101 / 200 [ 50%]  (Sampling)
#> Chain 1: Iteration: 120 / 200 [ 60%]  (Sampling)
#> Chain 1: Iteration: 140 / 200 [ 70%]  (Sampling)
#> Chain 1: Iteration: 160 / 200 [ 80%]  (Sampling)
#> Chain 1: Iteration: 180 / 200 [ 90%]  (Sampling)
#> Chain 1: Iteration: 200 / 200 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 5.0462 seconds (Warm-up)
#> Chain 1:                3.53858 seconds (Sampling)
#> Chain 1:                8.58478 seconds (Total)
#> Chain 1: 

#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

draws_stan <- popMethods::popReconstruct_posterior_draws(
  fit = fit_stan,
  inputs = demCore::burkina_faso_initial_estimates,
  settings = settings,
  value_col = "value",
  software = "stan",
  method_name = "original"
)

summary_stan <- popMethods::popReconstruct_summarize_draws(
  draws = draws_stan
)