Calculate Hyperparameters for the popReconstruct Model

Use elicited expert opinion about the observable marginal quantities in the popReconstruct model to quantify uncertainty in the input parameters.

Elicitation statement is "there is a \(p_{\nu}\)% probability that the true values are within plus-or-minus \(\eta_{\nu}\) x 100 percent of the initial point estimates. See references for more details.

popReconstruct_hyperparameters(abs_deviations, p, alpha, survival = NULL)

Arguments

abs_deviations
[list()]
Named list of absolute deviations \(\eta_{\nu}\) used in elicitation statement for each ccmpp/model component.

p
[list()]
Named list of central \(p_{\nu}\)% probability interval used in elicitation statement for each ccmpp/model component.

alpha
[list()]
Named list for alpha (shape) parameter of inverse gamma distribution for each ccmpp/model component.

survival
[numeric()]
Initial age-sex specific survival proportion estimates. Must be non-NULL if 'survival' parameter is included in the other input arguments.

Value

Nested list of alpha (shape) and beta (scale) parameters for the variance of each model component.

References

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2013. “Reconstructing Past Populations With Uncertainty From Fragmentary Data.” Journal of the American Statistical Association 108 (501): 96–110. https://doi.org/10.1080/01621459.2012.737729.

popReconstruct R Package

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal Statistical Society. Series A: Statistics in Society 178 (4): 977–1007. https://doi.org/10.1111/rssa.12104.

markalava/Bayesian-Reconstruction github repo

See also

Examples

abs_deviations <- list(
  srb = 0.1,
  asfr = 0.1,
  population = 0.1,
  survival = 0.1,
  net_migration = 0.2
)
p <- list(
  srb = 0.9,
  asfr = 0.9,
  population = 0.9,
  survival = 0.9,
  net_migration = 0.9
)
alpha <- list(
  srb = 0.5,
  asfr = 0.5,
  population = 0.5,
  survival = 0.5,
  net_migration = 0.5
)
hyperparameters <- popReconstruct_hyperparameters(
  abs_deviations,
  p,
  alpha,
  survival = demCore::thailand_initial_estimates$survival$value
)